Server_Django/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-jwt
pycrypto
